requirements.txt
matplotlib>=3.8.2
numpy>=1.26.4
openslide_python>=1.3.1
pandas>=2.2.0
scipy>=1.12.0
skimage>=0.0
torch>=2.0
torchvision>=0.16.2
tqdm==4.66.1